Move EqualizationParameters to new files
